Create tic_tac_toe.py
This was the first milestone project for the 2022 Complete Python Bootcamp from Zero to Hero on Udemy. I wanted to work it out without watching any of the guided or step-by-step parts where he walks you through it, so my solution is different from what they would have expected. 
I didn't like the idea of asking people to choose a square by number without defining what the numbers were, so mine works a little differently from theirs; but it does work.
I know there are better, more elegant ways to achieve this; but I'm learning and I want to remember the places I pass through on the journey.